fastapi_backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
opencv-python
opencv-python-headless
python-multipart
